+++ src/senaite/core/browser/controlpanel/manufacturers/instruments.py
@@ -24,6 +24,10 @@ from bika.lims.controlpanel.bika_instruments import InstrumentsView
 
 class ManufacturerInstrumentsView(InstrumentsView):
 
+    def __init__(self, context, request):
+        super(ManufacturerInstrumentsView, self).__init__(context, request)
+        self.context_actions = {}
+
     def isItemAllowed(self, obj):
         obj = api.get_object(obj)
         uid = obj.getRawManufacturer()